requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
Pillow
python-dotenv
google-generativeai
langchain
faiss-cpu
xlsxwriter
openpyxl